feat: Likes model created
Was deleted the raiting field from posts and instead was created the likes model.
This model will be the relationshio between the posts and users tables, a user has
many likes, this means that a posts can like it to a user, and post has many likes.

This is a many to many association.