feat: Password template
Password template created. This template will render the form with
the only field to update the password, this field is separated from the
form to update the user info that's why the password field was removed
from the form to update the user info.
Also was remove the form rendered in the 'app/views/uers/edit.html.erb' template