fix: Object event appended to the tag
In the previous changes, the appendToCombobox function
evaluated if the value exists, if true, then pass the
value as id and as value, otherwise it will be passed the
tag text from the list.

The problem was, due to the value param comes from another
function, the appendToCombobox function receives the param
as an array, so, the value always exists and an "object event"
was passed as the text for the tag.

The solution was evaluate if the typeof value == "string", if true,
this means that this text is a new tag, otherwise the text from the
list of tags is passed