Email confirmation
Added to Gemfile the jwt and rails_refactor gems.

The rails_refactor gem was added to the project to rename
a controller.

Token service

- Was created the functionality to encode and decode tokens using JWT.
  This functionality was created as a service
- After a user register to the system, a new token is generated with the
  user_id and user email, after that the email is sent

Was added a default_url_options config into the development environment
configurations. This defines a host and we can use in the mailer views
the url prefix